Replace Twisted API by subprocess.Popen
Twisted is a powerful framework but the fact that it enforces a
asynchronous work-flow and that it comes with its list of dependencies
which are not trivial to get it right (e.g. on Windows) makes it more
difficult to integrate. The benefit from Twisted is a fine-grain and
efficient process communication API. By removing this dependency, we
loose the capability of displaying accurate progress during the tool
execution but saves us from an (yet) overcomplicated build environment.

Using subprocess.Popen ensure maximum compatibility with Python and
avoid an additional dependency to GPS.

* share/gnathub/lib/GNAThub/__init__.py:
    Replace Twisted API calls by subprocess.Popen use. Re-organize the
    GNAThub.Plugin interface by removing obsolete methods and
    re-integrating it into a synchronous work-flow.

* share/gnathub/core/*.py:
    Update core plugins following API changes.

* src/gnathub-main.adb:
    Minor log format update for consistency with the core plugins
    output.

Change-Id: I4afb3126ad4a0c46509227780487939340a7dcfb
TN: MA07-064