Repository re-organization.
The goal is to support both the developer's mode using a virtual
Python environment and the continuous integration / production
environment driven by Anod.

First, remove the checked in SQLAlchemy sources and use Python
virtualenv to create the environment and download the necessary
packages.

Then create a setup.py for the GNAThub Python library (that is
implemented in both Python and Ada) to install it with Python's
setuptools like a standard project.

Change-Id: I48217ed6c789ee96a7e220849f909682cc6b8621
TN: MC17-038