Tentative fix for GNATmetric test.
The testcase tries to call sqlite3 command-line tool which is not
available by default on all machines/platforms. Since we don't want to
maintain an install of this tool nor handle it as part of GNAThub's
dependencies, this patch implements a Python wrapper on top of
SQLAlchemy that can query a SQLite database to replace the need for
sqlite3.

t/tests/gnatmetric/ChangeLog:

 * gnathub_query.py: new file to replace sqlite3.
 * test.sh: update the test file to use gnathub_query.py instead of
   sqlite3.

Change-Id: I94330671cb3216b8a8508ee72b3344d43d1e5090
TN: N113-034