DSH-22 Fix run of sonar-runner
The java code requires all resources to be associated with projects.

The Resource() method was used by plugins to create resources,
sometimes not belonging to a project (for instance, in the case
of .adb.gcov files being generated by gcov but not belonging to
the project tree).

To fix this, implement a method Resource.get(), which only
returns resources that do exist, and use this in the plugins.

Change-Id: I215796b39a73cbd7ed63ca0e45138a2746201343