gnathub: Fix SonarRunner execution failures.
The sonarconfig.py plugin creates the sonar-runner.properties file which
describes the project to the Sonar Runner. So far, the sonar.sources
property was hard-coded to "." ie. the current directory.

This led to plethora of invalid scenarios when gnathub is not invoked
from the same directory as the project file it is fed with (even running
gnathub from the same directory cannot guarantee a successful
execution).

This patch fixes this issue by setting the sonar.sources property to a
comma-separated list of Source_Dirs as declared in the project file.

src/ChangeLog:

    * gnathub/gnathub-project.ad[bs] (Source_Dirs): new routine
    * gnathub/gnathub-python.adb: define a new Python static method for
      the Project class: source_dirs()
    * lib/GNAThub/__init__.py (Project.source_dirs): new routine

share/gnathub/core/ChangeLog:

    * sonarconfig.py: set sonar.sources property to a comma-separated
      list of source directories

testsuite/tests/ChangeLog:

    * api-python/project/*: test availability of the
      Project.source_dirs() static method

Issue: DSH-31
Change-Id: Id108129ab327453a381def79ed5c8d3644afe4bc