Refactor logging module.
Distinguish routines to print GNAThub output with routines to log the
execution of GNAThub and its plug-ins.

Define a new Logger class and export it to Python. Plug-ins can simply
use the pre-defined self.log instance variable for logging.

Update all occurrences of printing and logging in the sources.
Also update two tests to ensure they do not attempt to run any other
plug-in than SonarConfig.

Change-Id: I47bd3a4fc243a70c12138ca2b8e854d770dbae82