Rewrite the sonar-ada-plugin.
The previous version was mostly using deprecated APIs, which prevented
us from using most of the recent features of SonarQube. It was also not
following SonarQube plug-ins guideline, which could give a bad
impression to users.

The new version leverages the most of the new APIs, but still keeps
compatibility with the current version of GNAThub.

Among other things, here what changed:

  * Source import is now completely handled by Sonar
    * This means that source scanning is a lot faster and that the
      integration with the SonarQube UI is far better

  * Deprecated API calls have been (almost) all replaced by the newest
    API calls
    * This ensures that we will stay compatible with future versions of
      SonarQube.

  * The SonarAdaPlugin can now be used in multiple-languages and
    multiple-modules projects.

Change-Id: Ib88504f59539190b073d5db28680ea4fb874ffb2