gnathub: implement custom logging.Handler
This handler relies on GNATCOLL.Traces and the exposed API from GNAThub
to log its messages. That way, we can benefit from the very powerful
logging module in Python scripts and plug-ins. It also integrates nicely
with existing frameworks that already use the usual Python way for
logging.

Change-Id: Ifd71c1e1d3507ea04d24b52eac9a9ce5ab18e224
Issue: DSH-37