Preliminary support for bulk message insertion
* db.model:
Remove intermediary tables lines and lines_messages: all messages are now
in resources_messages, with line = 0 meaning a message not associated
to a line but to the resource itself.

* gnathub-python-database.adb:
Add support for a command Resource.add_messages, used for bulk message
insertion. This uses a prepared statement and a manual transaction
for efficiency.
Simplify implementation of Resource.add_message with the new database
schema.

* __init__.py:
Resource.add_message: Change default value for line, more in line with the
schema.
Document Resource.add_messages.

Regenerate the ORM.

For N805-033.
Change-Id: Ib611c52852fbf8b40b4e81636fddbb0e54f0d4e5