Copy all source files to a local cache directory.
This is a workaround to SonarQube SourceImporter which looks recursively
in all source directories whereas the semantic of GPR files is to only
look for files in source directories.

It also generates the appropriate mapping file between original source
files and copies. This mapping file is then used by the SonarAdaPlugin
to convert path stored in the database into path pointing to the local
cache.

Change-Id: I24943975f31628003281be4699167efb78147eae