Use -U more consistently across tools
Passing -U to GNAT tools (such as GNATmetric, GNATcheck, ...) force the
processing of the entire closure of source of the input project. This is
relevant in the context of GNAThub where we want to generate metrics for
as many sources as possible (and possibly filter some of these out at a
later point).

This patch specify the missing -U to the GNATcheck tool.

Change-Id: Ibb71243f115775797c82a62f7dd11aefe56ad336
TN: NB20-021