Add Makefile target to set version
Use the content of SONAR_ADA_PLUGIN_VERSION to set the version, eg.:

    $ SONAR_ADA_PLUGIN_VERSION=x.y.zw make set-version

This is used by sonar-ada-plugin.anod to compute a version containing
the build date (eg. 1.1.0w-20150222) so that it provides an easy way to
trace a plugin loaded in SonarQube to its date of build (the exact
version is displayed in the Update Center).

Change-Id: Iccdc9e146cb92a82230b9ca81322e8d431bf1986
TN: NA24-021