Add support for CodePeer suppressed messages
Map them to 'info' messages.
Add INFO as a possible value for the CodePeerSeverity enum.

The codepeer .csv report overrides the category of the rule
for suppressed messages with the string "suppressed". This means that
we need to create a "suppressed" variant of each rule in the codepeer.xml
rule file: modify the script which generates this file accordingly.

Regenerate codepeer.xml with a recent codepeer.

For P928-021.

Change-Id: I5ce5099987948afe09608d204303c8f47c987879