fix(html-report): ensure that we can encode files as UTF-8
The HTML report module output JSON files that are UTF-8 encoded. In
order for this serialization process to work, the input files need to be
decoded first.

For now, the implementation expects all input files to be UTF-8
otherwise the source code won't be available in the HTML report.

In the future GNAThub could expose a switch to allow the user to specify
the input encoding.

Change-Id: If9ae4bec08ca15301caab447665346e7da94d5f7