Fix webui build process
Use es6 target for typescript
Use proper node modules versions
Remove dll budles plugin: this plugin was a shortcut
to generate vendor bundle + reference it in the bundled app,
but in fact it was forcing the vendor entry point name to "vendor"
when it should have been the same as the name defined in
the bundle-manifest.json: vendor_lib.
Hence the 'vendor_lib' not defined.

So make it work, do:

> npm install
> npm run build:dev
> npm run server:dev
  (after having put the data dir content in /dist/data)

And it requires at least node 8.0.0

TO-DO: the final js files are a bit under 2M, some minification
might be good to reduce it a bit, but as it's a loaded locally,
it's no big deal.

Part of S913-015

Change-Id: I9c37f86e54adeed938f684f5a8e23646e93c0291