WebUI: Put version control for Codepeer
Gnathub now create a file version.txt with the codepeer version in it.
While starting the server, the codepeer-passed/ route is called
and then verify if codepeer executable exist
and if its version is the same as the one registered.
If everything is correct, return 200.
If version are not matching, return 202,
which trigger an error on the interface.

Change-Id: Ib25ac3d0fae5cd1268a879d67ca7d6bde7f57921
For: SB26-048