U723-025 Reorganize Python files to support virtualenvs
A Python virtualenv might overwrite the site-packages entry
in sys.path; when this happens, plugin-runner.py can no longer
load GNAThub.py.

To address this, place GNAThub.py in the core directory rather
than in the Python install.

This means that tests can no longer simply rely on the GNAThub
module being available for testing - to make this available
to tests, set an environment variable COREDIR to point to the
directory containing this. (We could make this simpler by setting
the PYTHONPATH for all tests, but this presents the risk of
hiding setup problems where the python-runner.py would normally
not find the GNAThub module)

Simplify the structure in the process, removing src/lib and
placing all Python files under share/gnathub/core.

Simplify the Makefile accordingly.

Remove use of deprecated API in the testsuite driver.

Change-Id: I2574afd9560eba7ff9a6e0a98836d1a04b11c807